(#39) Build: pass the ProgressIndicator correctly
