+++ core/src/main/resources/messages/HaskeletorBundle.properties
@@ -23,6 +23,3 @@ progress.installing-tool.text={0}: {1} / {2} packages
 progress.installing-tool.in-progress=Processing: {0} {1,choice,0#|0<and {1} other}
 
 workflow.build-project-with-dependencies.title=Building project, starting REPL(s) and preloading cache
-
-workflow-step.build-project-dependencies.text=Busy building project's dependencies
-workflow-step.build-project.text=Busy building project
